* Fixing #334 machine now sleeps during encoding (thanks to Don Gafford) * Fixing After Conversion command running after every encoding (thanks to Don Gafford)
